Move hfp and route queries to App.
